requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
binaryornot==0.4.4
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cookiecutter==2.5.0
cron-descriptor==1.4.0
cryptography==41.0.5
daphne==4.0.0
Django==4.2.7
django-browser-reload==1.12.1
django-celery-beat==2.5.0
django-dynamic-preferences==1.16.0
django-polymorphic==3.1.0
django-tailwind==3.6.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
docopt==0.6.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
kombu==5.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
msgpack==1.0.7
persisting-theory==1.0
prompt-toolkit==3.0.41
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==23.3.0
pyq3serverlist==0.3.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
python-valve==0.2.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rich==13.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
Twisted==23.10.0
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
zope.interface==6.1
